error handling property info functions

+++ src/routes/property_information.py
@@ -3,39 +3,42 @@ import json
 from flask import Blueprint, request
 
 from src.config.logger import logger
-from src.db_config import property_information
+from src.db_config import property_information, user
 from src.utilities.respond import success
 from src.utilities.exceptions.exceptionfactory import ExceptionFactory
+from src.utilities.schemahandler import SchemaHandler
 
 information_bp = Blueprint(name="information", import_name=__name__)
-
+schema_handler = SchemaHandler()
 
 @information_bp.route('/information/<string:property_id>', methods=['GET'])
 def get_information(user_id: str, property_id: str):
     """Gets an individual property information."""
-    try:
-        logger.info(f"UserId: {user_id}")
-        logger.info(f"PropertyId: {property_id}")
-        data = property_information.find_one({'property_id': property_id,
-                                              'user_id': user_id})
+    logger.info(f"UserId: {user_id}")
+    logger.info(f"PropertyId: {property_id}")
+    data = property_information.find_one({'property_id': property_id,
+                                            'user_id': user_id})
+    logger.info(f"Data: {data}")
+    if data == None:
+        raise ExceptionFactory("Information not found").invalid_dict_parameter_value()
 
-        data["_id"] = str(data["_id"])
-
-        return {"success": True, "message": "information_data", "body": json.loads(data)}
-    except Exception as e:
-        logger.error(e)
-        return {"success": False, "message": "Error getting information data"}
+    data["_id"] = str(data["_id"])
+    return {"success": True, "message": "information_data", "body": json.loads(json.dumps(data))}
+   
 
 
 @information_bp.route('/information', methods=['POST'])
 def create_information(user_id: str):
     """Creates a property information."""
-    data = request.json
-    data["user_id"] = user_id
-    response = property_information.insert_one(data)
-    if not response.acknowledged:
-        raise ExceptionFactory("").database_operation_failed()
-    return success()
+    response = user.find_one({'user_id': user_id})
+    logger.info(f"Response: {response}")
+    if response == None:
+        raise ExceptionFactory("").database_operation_failed()                                      
+    else:
+        data = request.json
+        data["user_id"] = user_id
+        response = property_information.insert_one(data)
+        return success("Property information created")
 
 
 @information_bp.route('/information/<string:property_id>', methods=['PUT'])
@@ -43,16 +46,17 @@ def update_information(user_id: str, property_id: str):
     """Updates individual property information."""
     logger.info(f"UserId: {user_id}")
     logger.info(f"PropertyId: {property_id}")
-    try:
-        data = request.json
-        logger.info(f"Payload: {data}")
-        property_information.update_one({'user_id': user_id,
-                                         'property_id': property_id}, {'$set': data})
-        return {"success": True, "message": "information_data"}
+    data = request.json
+    query = {'user_id': user_id, 'property_id': property_id}
 
-    except Exception as e:
-        logger.error(f"Error: {e}")
-        return {"success": False, "message": "information_data"}
+    response = property_information.find_one(query)
+    if response == None:
+        raise ExceptionFactory("").database_operation_failed()
+    else:
+        response = property_information.update_one(query, {'$set': data})
+        logger.info(f"Response: {response}")
+        return success("information data updated")
+    
 
 
 @information_bp.route('/information/<string:property_id>', methods=['DELETE'])
@@ -60,10 +64,12 @@ def delete_information(user_id: str, property_id: str):
     """Deletes a property information."""
     logger.info(f"UserId: {user_id}")
     logger.info(f"PropertyId: {property_id}")
-    try:
+    query = {'user_id': user_id, 'property_id': property_id}
+    response = property_information.find_one(query)
+
+    if response == None:
+        raise ExceptionFactory("").database_operation_failed()
+    else:
         property_information.delete_one({'user_id': user_id,
-                                         'property_id': property_id})
-        return {"success": True, "message": "information_data"}
-    except Exception as e:
-        logger.error(f"Error: {e}")
-        return {"success": False, "message": "information_data"}
+                                            'property_id': property_id})
+        return success("information data deleted")
